src/Hikrobot/requirements.txt
numpy >= 1.21
pillow >= 9.0
grpcio-tools ~= 1.30
pytest ~= 5.4
grpc-interceptor ~= 0.15.4